[IMP+FIX] mgmtsystem_nonconformity: Migration scripts + several things
[ADD] Added migration scripts for 0.1 --> 1.0 in the modules that needed it.

mgmtsystem_nonconformity:

* o --> open c --> done state change.
* Migrating action relations to a many2many table.
* Concatening action comments in a single field

mgmtsystem_audit and mgmtsystem_review:

* o --> open c --> done state change.

[FIX] I had accidentally broken the nonconformity/action migration shortly before commit.

[FIX] Fixed "Unknown field state for domain" errors.

In 7.0, it's necessary, in a view, to have the field present in a view in order to use it
in a domain. I added an invisible "state" field in relevant views to fix the errors.

[CHG] Adjusted mgmgsystem_nonconformity to OpenERP 7.0.

Like with mgmtsystem_action, the problem comes from changes made in OpenERP's messaging system.

[FIX] Broken reference to wiki's group_id.

In OpenERP 7, document_page had a parent_id instead of wiki's group_id. I've fixed the broken references that we had left in nonconformity and audit.

[FIX] Fixed wrongful merge.

The merge of nc-extend-7.0 was incorrect. It wrongly used the 6.1 version of mgmtsystem_audit.xml and mgmtsystem_nonconformity/mgmtsystem_nonconformity.xml. Fixed this.

[CHG] Updated statusbar/buttons's layout to conform to OpenERP 7.0 style.

Instead of a group at the bottom of the form, it's now a <header> tag at the top.

Changed Actions, Audit, Nonconformity and Review.