Refactored the module mgmtsystem_hazard
Removed mgmtsystem_hazard_risk_* and moved in a submodule.
Splitted the file mgmtsystem_hazard/mgmtsystem_hazard_data.xml into
xml files per model.

I moved the risks modules in a different module because it creates a
problem during the loading of the model res_company.

During loading, the res_company tries to do an alter table and set null
the risk computation id. Has the computation id aren't yet loaded in
the database, it raises a constraint error because it can't find a
proper default value.