Added module to bring back workflows of version 8.0 in mgmtsystem_nonconformity
This restores the old behavior from version 8.0 where the quality was
ensured by a cycle of analyzing an NC and approving the result of the
analysis and using those results as the base for planning measures and
deciding on the detailed actions.
In addition, the workflow now supports a new stage.
After the work on the action plan is done, the NC cannot be closed
directly. The responsible user has to send his NC to the evaluation
stage, where the approving user evaluates the effectiveness of his work
on the NC, leaves feedback in the evaluation comments field and has to
signs his evaluation. After an approving user signs the evaluation, the
responsible user is able to close his case.