Fix odoo warning with duplicate field string
The enterprise module "quality_control" adds a field with the string
"Quality" in "res.config.settings", so does mgmtsystem.

Which results in a lot of warnings at the loading of the modules:

Two fields (module_quality_control, module_mgmtsystem_quality) of res.config.settings() have the same label: Quality.

I propose a slight change to the string to avoid this.